backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "goldpulse-api"
version = "0.1.0"
description = "GoldPulse FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "redis>=5.0",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
